Move templates from shared_templates to feature modules
Relocates LogInFormTemplate, RegisterFormTemplate,
ConfirmPasswordInputTemplate, NewTagFormTemplate, and
TransactionTableRow from shared_templates.rs to their respective feature
modules (log_in.rs, register_user.rs, tag.rs, transaction.rs).
