+++ TODO.md
@@ -1,6 +1,5 @@
 # To Do
 
-- Move non-shared code out of src/shared_templates.rs into the appropriate modules.
 - Add rules-based auto-tagger (new branch, merge back w/ PR):
   - Add a user-configurable rules-based (text patterns) auto-tagging feature
   - Use rules-based tagger for automatically tagging imports
@@ -8,6 +7,7 @@
 - Use rules-based tagger to mark internal transfers
 - Add thousands separator to monetary amounts by implementing custom currency filter for Askama
 - Align dashboard elements nicely
+- Change log in and registration pages to just ask for password
 - Split up `src/transaction.rs` into module `src/transaction/*.rs`
 - Ensure that simple and full csv imports from Kiwibank do not create duplicate
   transactions.
